update: Documentation update
Now I have updated documentation to state that this tools is.
It's quite safe and stuff.
Also updated from 0.1.0 => 0.9.0, now there might be mostly internal
changes but no longer that much api changes. I might make new one for
decrease and increase so it can be overridden that "lock" of not going
under 1% or over 100% if some people like that.

Also now I will write some tests for this too so, so I need to change
internals to point that those can be tested, quite lot of the things are
hardcoded what is bad.